Add dbcopy option to DesignDocument.MapReduce
Refactored DesignDocument to a single class.
Added dbcopy option to DesignDocument.MapReduce.
Added dbcopy test.

BugId: 50046
